Add UI back into same repo as service code
This will allow for an easier build and deployment structure
going forward. In future updates, when we are ready to create
distributions we can build a Docker image and not have to try
to bring in multiple repos.

I have kept the sonar projects separate so the build will run
sonar for service only and then for ui only since they don't
get built all through Maven.

Service (maven/java) code lives in the directory service and
UI (node/vue/quasar) code lives in the directory ui.

Updated the github actions to account for new directories and
build processes for both "modules".